backend/requirements.txt
django==2.2.16
pytest-django==3.8.0
django-filter==21.1
pytest-pythonpath==0.7.3
pytest==5.3.5
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
mixer==7.1.2
Faker==12.0.1
djoser~=2.1.0
djangorestframework-simplejwt==4.7.2
djangorestframework~=3.12.4
asgiref==3.2.10
PyJWT==2.1.0
drf-yasg
psycopg2-binary==2.9.5
python-dotenv
pytz==2020.1
sqlparse==0.3.1
Pillow==9.3.0
reportlab
gunicorn==20.0.4
